Moved project creation to inside CurrentOrganization action and changed the way projects are fetch in project Root view
