+++ CompositeUi/src/actions/Projects.js
@@ -8,14 +8,10 @@
  * file that was distributed with this source code.
  */
 
-import {routeActions} from 'react-router-redux';
-
-import {routes} from './../Routes';
-
 import ActionTypes from './../constants/ActionTypes';
 
 import ProjectsQueryRequest from './../api/graphql/query/ProjectsQueryRequest';
-import CreateProjectMutationRequest from './../api/graphql/mutation/CreateProjectMutationRequest';
+
 import TaskManagerGraphQl from './../api/graphql/TaskManagerGraphQl';
 
 const Actions = {
@@ -29,34 +25,6 @@ const Actions = {
         dispatch({
           type: ActionTypes.PROJECTS_RECEIVED,
           projects: data.response.projects.edges
-        });
-      });
-  },
-  createProject: (projectInputData) => (dispatch) => {
-    dispatch({
-      type: ActionTypes.PROJECTS_CREATING
-    });
-    const mutation = CreateProjectMutationRequest.build(projectInputData);
-
-    TaskManagerGraphQl.mutation(mutation, dispatch);
-    mutation
-      .then(data => {
-        const project = data.response.createProject.project;
-
-        dispatch({
-          type: ActionTypes.PROJECTS_CREATED,
-          project,
-        });
-        dispatch(
-          routeActions.push(routes.project.show(project.organization.slug, project.slug))
-        );
-      })
-      .catch((response) => {
-        response.then((errors) => {
-          dispatch({
-            type: ActionTypes.PROJECTS_CREATE_ERROR,
-            errors
-          });
         });
       });
   }
