+++ CompositeUi/src/constants/ActionTypes.js
@@ -50,12 +50,11 @@ const ActionTypes = {
   PROJECTS_FETCH_ERROR: 'PROJECTS_FETCH_ERROR',
   PROJECTS_RECEIVED: 'PROJECTS_RECEIVED',
 
-  PROJECTS_CREATING: 'PROJECTS_CREATING',
-  PROJECTS_CREATED: 'PROJECTS_CREATED',
-  PROJECTS_CREATE_ERROR: 'PROJECTS_CREATE_ERROR',
-
   CURRENT_ORGANIZATION_FETCHING: 'CURRENT_ORGANIZATION_FETCHING',
   CURRENT_ORGANIZATION_RECEIVED: 'CURRENT_ORGANIZATION_RECEIVED',
+  CURRENT_ORGANIZATION_PROJECT_CREATING: 'CURRENT_ORGANIZATION_PROJECT_CREATING',
+  CURRENT_ORGANIZATION_PROJECT_CREATED: 'CURRENT_ORGANIZATION_PROJECT_CREATED',
+  CURRENT_ORGANIZATION_PROJECT_CREATE_ERROR: 'CURRENT_ORGANIZATION_PROJECT_CREATE_ERROR',
 
   CURRENT_PROJECT_FETCHING: 'CURRENT_PROJECT_FETCHING',
   CURRENT_PROJECT_RECEIVED: 'CURRENT_PROJECT_RECEIVED',
