+++ CompositeUi/src/reducers/Projects.js
@@ -27,12 +27,6 @@ export default function reducer(state = initialState, action = {}) {
     case ActionTypes.PROJECTS_FETCH_ERROR:
       return {...state, error: true};
 
-    case ActionTypes.PROJECTS_CREATED:
-      return {...state, projects: [...state.projects, action.project]};
-
-    case ActionTypes.PROJECTS_CREATE_ERROR:
-      return {...state, errors: action.errors};
-
     default:
       return state;
   }
